falco-ai-security/backend/requirements.txt
# FastAPI和Web服务
fastapi==0.104.1
uvicorn[standard]==0.24.0
pydantic==2.5.0
pydantic-settings==2.1.0

# HTTP客户端
httpx==0.25.2
requests==2.31.0

# 数据库连接
neo4j==5.15.0
redis==5.0.1

# AI和机器学习
openai==1.3.7
pinecone-client==2.2.4
numpy==1.24.3
scipy==1.11.4

# 数据处理
pandas==2.1.4
json5==0.9.14
pyyaml==6.0.1

# 日志和监控
loguru==0.7.2
psutil==5.9.6
watchdog==3.0.0

# 异步处理
aiofiles==23.2.1
aioredis==2.0.1
celery==5.3.4

# WebSocket和实时通信
websockets==11.0.3

# 安全和认证
passlib[bcrypt]==1.7.4
python-jose[cryptography]==3.3.0
python-multipart==0.0.6

# 时间处理
python-dateutil==2.8.2
pytz==2023.3

# 环境变量
python-dotenv==1.0.0

# 图数据处理
networkx==3.2.1

# JSON处理
orjson==3.9.10

# 文件处理
chardet==5.2.0

# 系统监控
psutil==5.9.6

# 测试框架
pytest==7.4.3
pytest-asyncio==0.21.1
pytest-cov==4.1.0

# 开发工具
black==23.11.0
flake8==6.1.0
isort==5.12.0
mypy==1.7.1

# 文档生成
mkdocs==1.5.3
mkdocs-material==9.4.8